+++ src/components/show-more-btn.js
@@ -1,0 +1,9 @@
+/**
+ * Returns show more button html
+ * @return {string}
+ */
+const getBtnShowMoreHtml = () => (
+  `<button class="films-list__show-more">Show more</button>`
+);
+
+export {getBtnShowMoreHtml};
