+++ webpack.config.js
@@ -9,7 +9,7 @@ module.exports = {
   },
   devtool: `source-map`,
   devServer: {
-    contentBase: path.join(__dirname, `public`),
+    contentBase: path.join(__dirname, `/public/`),
     watchContentBase: true
   }
 };
